# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Using a translation service as an example",
        "title": "Go Clean Template API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/product/add-product": {
            "post": {
                "description": "Add a product to inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Product",
                "operationId": "add-product",
                "parameters": [
                    {
                        "description": "Enter product details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.addProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/product/get-product": {
            "get": {
                "description": "Get one product details from inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Product",
                "operationId": "get-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "01F8Z4P8KQ4P7CSHMSZETEATWX"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "MacBook Pro 9999"
                },
                "id": {
                    "type": "string",
                    "example": "01HV73P3B06PTPXEMYCCBMM03C"
                },
                "name": {
                    "type": "string",
                    "example": "Laptop"
                },
                "unit_price": {
                    "type": "number",
                    "example": 99.9
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "v1.addProductRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "01F8Z4P8KQ4P7CSHMSZETEATWX"
                },
                "description": {
                    "type": "string",
                    "example": "MacBook Pro 9999"
                },
                "name": {
                    "type": "string",
                    "example": "Laptop"
                },
                "unit_price": {
                    "type": "number",
                    "example": 99.9
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}